Feat: book implementations

+++ backend/src/services/books.service.ts
@@ -1,4 +1,6 @@
+import { BooksDto } from '@/dtos/books.dto';
 import { Books } from '@/entity/books.entity';
+import { HttpException } from '@/exceptions/httpException';
 import { BooksRepository } from '@/repository/books.repository';
 import Container, { Service } from 'typedi';
 
@@ -11,10 +13,15 @@ export class BooksService {
   }
 
   async findOne(id: number): Promise<Books | undefined> {
-    return await this.booksRepository.findOne(id);
+    const book = await this.booksRepository.findOne(id);
+    if (!book) throw new HttpException(404, 'Book not found');
+
+    return book;
   }
 
-  async create(book: Books): Promise<Books> {
+  async create(book: BooksDto): Promise<Books> {
+    const existingBook = await this.booksRepository.findByTitle(book.title);
+    if (existingBook) throw new HttpException(403, 'Book title already exists');
     return await this.booksRepository.create(book);
   }
 
@@ -22,7 +29,10 @@ export class BooksService {
     return await this.booksRepository.update(id, book);
   }
 
-  async delete(id: number): Promise<void> {
+  async delete(id: number): Promise<Books> {
+    const book = await this.booksRepository.findOne(id);
+    if (!book) throw new HttpException(404, 'Book not found');
     await this.booksRepository.delete(id);
+    return book;
   }
 }
